feat: Update auth actions and components
The commit includes changes to the auth actions and components. The changes include adding new imports, dispatching new actions, and updating button text in the LoginScreen component. Additionally, there are updates to the CreateLabourModal, Docent, EditLabourModal, Header, ProSidebar components as well as the useGetEducators hook. Finally, there are modifications to the educatorSlice, labourSlice, and periodSlice files to clear state data.